Fix various config file reading bugs
If a key=value pair contains a space (such as "SNDFORMAT0=8-bit Signed"
then scanf() stops at the space.  Parsing continues after the space,
which usually fails and the rest of the file is discarded.  Fix this by
reading in entire lines at a time.

If the parser finds 10 characters that looks like a key, it will try to
match it against the known keys.  However if the 10 characters are not
one of the known keys, then it will walk off the end of the array and
segfault.