Use SPDX tags for copyrights and license grants
From COPYING, Section 9:  "If the Program does not specify a version
number of this License, you may choose any version ever published by the
Free Software Foundation."  None of the license grants specify a
version, so we use the "GPL-1.0-or-later" license identifier (for lack
of a GPL-any-version alternative) even though COPYING is the verbatim
GPL 2.0 license text.  See discussion and examples at
https://wiki.spdx.org/view/Legal_Team/only-operator-proposal

I'm adding myself as a copyright holder, with 2014 as the starting year
according to the Git history.  I'm also licensing the autotools input
files (configure.ac, Makefile.am) under the same license as the other
existing source files, as described above.