Change license to GPL-2.0-or-later
See fb32ce746e429bc1454bacfa10628832ed2f5cde for additional context.
The original sources never specified a version, but they did ship with
the GPLv2 license text which indicates that any GPL version applies in
this situation.

But since then, I've completely rewritten the code, and I want the
license to match that of tuxnes.